Added extra features
