+++ src/main/java/com/zipcodewilmington/scientificcalculator/CubeFeatures.java
@@ -1,0 +1,10 @@
+package com.zipcodewilmington.scientificcalculator;
+
+public class CubeFeatures {
+    private double result;
+    public double cube(double number1) {
+        result = number1 * number1 * number1;
+        return result;
+    }
+
+}
