+++ src/main/java/com/zipcodewilmington/scientificcalculator/CubeRootFeatures.java
@@ -1,0 +1,12 @@
+package com.zipcodewilmington.scientificcalculator;
+
+public class CubeRootFeatures {
+
+    private double result;
+
+    public double cubeRoot(double num1) {
+        result = Math.cbrt(num1);
+        return result;
+    }
+
+}
