+++ src/main/java/com/zipcodewilmington/scientificcalculator/ScientificFeatures.java
@@ -100,15 +100,8 @@ public class ScientificFeatures {
         return result;
     }
 
-    public double cubeRoot(double num1) {
-        result = Math.cbrt(num1);
-        return result;
-    }
 
-    public double cube(double number1) {
-        result = number1 * number1 * number1;
-        return result;
-    }
+
 
 
 }
